+++ tests/testthat/test_rbindCommon.R
@@ -1,8 +1,8 @@
 
 
 test_that("Malformed inputs", {
-  #rbind.common(c(a = 1, b = 3), c(a = 3, b = 6))
-  #rbind.common(c(a = 1, b = 3), mtcars)
+  expect_error(rbind_compact(c(a = 1, b = 3), c(a = 3, b = 6)), "All inputs to rbind_compact must be data.frames.")
+  expect_error(rbind_compact(c(a = 1, b = 3), mtcars), "All inputs to rbind_compact must be data.frames.")
 })
 
 ###
@@ -13,14 +13,14 @@ test_that("rbind data.frames", {
   df2 <- data.frame(d = 100, a = 11:13)
   df3 <- data.frame(d = 1000, x = 101:103)
 
-  out <- rbind.common(df1, df2)
+  out <- rbind_compact(df1, df2)
   expect_equal(out, data.frame(a = c(1:3, 11:13)))
 
-  expect_warning(out2 <- rbind.common(df1, df2, df3),
+  expect_warning(out2 <- rbind_compact(df1, df2, df3),
                  "No common column names found.")
   expect_equal(out2, NULL)
 
-  out3 <- rbind.common(df1, df1b, df1c)
+  out3 <- rbind_compact(df1, df1b, df1c)
   expect_equal(out3, data.frame(a = 1:10, b = TRUE))
 
 })